--- a/lib/graphql/language/lexer.rb
+++ b/lib/graphql/language/lexer.rb
@@ -166,7 +166,7 @@ module GraphQL
 
       # Replace any escaped unicode or whitespace with the _actual_ characters
       # To avoid allocating more strings, this modifies the string passed into it
-      def replace_escaped_characters_in_place(raw_string)
+      def self.replace_escaped_characters_in_place(raw_string)
         raw_string.gsub!(ESCAPES, ESCAPES_REPLACE)
         raw_string.gsub!(UTF_8) do |_matched_str|
           codepoint_1 = ($1 || $2).to_i(16)
@@ -231,7 +231,7 @@ module GraphQL
         if !value.valid_encoding? || !value.match?(VALID_STRING)
           emit(:BAD_UNICODE_ESCAPE, ts, te, value)
         else
-          replace_escaped_characters_in_place(value)
+          self.class.replace_escaped_characters_in_place(value)
 
           if !value.valid_encoding?
             emit(:BAD_UNICODE_ESCAPE, ts, te, value)